update `OnboardingCountryPage` to use `countriesProvider`
